views: refactor templates
