+++ resource.php
@@ -67,11 +67,8 @@ if($_SERVER['REQUEST_METHOD'] === "GET" || $_SERVER['REQUEST_METHOD'] === "HEAD"
       $resource = $g->toRdfPhp();
 
       $scripts = get_values($resource, 'view:script');
-      include 'views/top.php';
-      include 'views/nav.php';
-      include 'views/'.view_router($resource).'.php';
-      // include 'views/nav.php';
-      include 'views/end.php';
+      $includes = array(view_router($resource).'.php');
+      include 'views/page_template.php';
     }
   }catch(Exception $e){
     var_dump($e);
