+++ views/end.php
@@ -1,4 +1,3 @@
-  </main>
   <?if(isset($scripts)):?>
     <?foreach($scripts as $script):?>
       <script src="<?=$script?>"></script>
