+++ views/travel.css
@@ -1,3 +1,4 @@
+@import url('base.css'); /* this is a hack */
 .map {
   text-align: center;
   overflow: hidden;
